requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
browser-use==0.1.0
playwright==1.40.0
langchain-openai==0.0.2
pydantic==2.5.0
python-dotenv==1.0.0
websockets==12.0
python-multipart==0.0.6
